backend/pyproject.toml
[project]
name = "agent-chat-ui-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.22",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "langgraph-cli[inmem]>=0.4.3",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.12",
]
